# docs/swagger.yaml
basePath: /
definitions:
  common.ErrorResponse:
    properties:
      details:
        additionalProperties: true
        type: object
      error_code:
        type: string
      message:
        type: string
    type: object
  db.ConcurrentJobLimit:
    properties:
      concurrent_jobs:
        type: integer
      is_default:
        type: boolean
      username:
        type: string
    type: object
  db.ConcurrentJobLimitUpdate:
    properties:
      concurrent_jobs:
        type: integer
    type: object
  db.ConcurrentJobLimits:
    properties:
      limits:
        items:
          $ref: '#/definitions/db.ConcurrentJobLimit'
        type: array
    type: object
  db.DeletionResponse:
    properties:
      id:
        type: string
    type: object
...
  db.NewQuickLaunchUserDefault:
    properties:
      app_id:
        type: string
      quick_launch_id:
        type: string
    type: object
...
  db.QuickLaunchUserDefault:
    properties:
      app_id:
        type: string
      id:
        type: string
      quick_launch_id:
        type: string
      user:
        type: string
    type: object
...
  db.UpdateQuickLaunchUserDefaultRequest:
    properties:
      app_id:
        type: string
      quick_launch_id:
        type: string
    type: object
...
host: localhost:60000
info:
  contact: {}
  description: API for the Discovery Environment's analyses, including Quick Launches
    and settings.
  license:
    name: 3-Clause BSD License
    url: https://github.com/cyverse-de/analyses?tab=License-1-ov-file#readme
  title: Analyses API
  version: "1.0"
paths:
...
  /quicklaunch/defaults/user:
    get:
      description: Get all of the Quick Launch user defaults for the logged in user.
      parameters:
      - description: Username
        in: query
        name: user
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/db.QuickLaunchUserDefault'
            type: array
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/common.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/common.ErrorResponse'
      summary: Get all Quick Launch user defaults
      tags:
      - quicklaunch-user-defaults
    post:
      consumes:
      - application/json
      description: Add a Quick Launch user default. A new UUID will be assigned to
        the user default and will be returned in the response.
      parameters:
      - description: Username
        in: query
        name: user
        required: true
        type: string
      - description: User default to add
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/db.NewQuickLaunchUserDefault'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/db.QuickLaunchUserDefault'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/common.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/common.ErrorResponse'
      summary: Add a Quick Launch user default
      tags:
      - quicklaunch-user-defaults
  /quicklaunch/defaults/user/{id}:
    delete:
      description: Delete the Quick Launch user default.
      parameters:
      - description: User Default ID
        in: path
        name: id
        required: true
        type: string
      - description: Username
        in: query
        name: user
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/db.DeletionResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/common.ErrorResponse'
      summary: Delete a Quick Launch user default
      tags:
      - quicklaunch-user-defaults
    get:
      description: Get a Quick Launch user default.
      parameters:
      - description: User Default ID
        in: path
        name: id
        required: true
        type: string
      - description: Username
        in: query
        name: user
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/db.QuickLaunchUserDefault'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/common.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/common.ErrorResponse'
      summary: Get a Quick Launch user default
      tags:
      - quicklaunch-user-defaults
    patch:
      consumes:
      - application/json
      description: Modifies an existing Quick Launch user default.
      parameters:
      - description: User Default ID
        in: path
        name: id
        required: true
        type: string
      - description: Username
        in: query
        name: user
        required: true
        type: string
      - description: Fields to update
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/db.UpdateQuickLaunchUserDefaultRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/db.QuickLaunchUserDefault'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/common.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/common.ErrorResponse'
      summary: Update a Quick Launch user default
      tags:
      - quicklaunch-user-defaults
...
  /settings/concurrent-job-limits:
    get:
      description: Lists all defined concurrent job limits.
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/db.ConcurrentJobLimits'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/common.ErrorResponse'
      summary: List concurrent job limits
      tags:
      - settings
  /settings/concurrent-job-limits/{username}:
    delete:
      description: Removes the explicitly set concurrent job limit for a user. This
        will effectively return the user's limit to whatever the default limit is.
      parameters:
      - description: Username
        in: path
        name: username
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/db.ConcurrentJobLimit'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/common.ErrorResponse'
      summary: Remove concurrent job limit
      tags:
      - settings
    get:
      description: Gets the concurrent job limit for a user. The returned job limit
        may either be the limit that was explicitly assigned to the user or the default
        job limit.
      parameters:
      - description: Username
        in: path
        name: username
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/db.ConcurrentJobLimit'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/common.ErrorResponse'
      summary: Get concurrent job limit
      tags:
      - settings
    put:
      consumes:
      - application/json
      description: Sets the concurrent job limit for a user. The user's limit will
        be updated explicitly in the database even if the requested limit is the same
        as the default limit.
      parameters:
      - description: Username
        in: path
        name: username
        required: true
        type: string
      - description: Job limit to set
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/db.ConcurrentJobLimitUpdate'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/db.ConcurrentJobLimit'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/common.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/common.ErrorResponse'
      summary: Set concurrent job limit
      tags:
      - settings
swagger: "2.0"